a bit more refactoring
